intial

+++ src/main/java/com/robson/pride/api/utils/ParticleUtils.java
@@ -74,7 +74,10 @@ public class ParticleUtils {
                     PrideVec3f vec = ArmatureUtils.getJointWithTranslation(renderer, living, AABB, joint);
                     if (vec != null) {
                         Vec3 delta = living.getDeltaMovement();
-                        renderer.level().addParticle(particle, vec.x(), vec.y(), vec.z(), ((new Random()).nextFloat() - 0.5F) * 0.02F + delta.x, (double) (((new Random()).nextFloat() - 0.5F) * 0.02F), ((new Random()).nextFloat() - 0.5F) * 0.02F + delta.z);
+                        renderer.level().addParticle(particle, vec.x(), vec.y(), vec.z(),
+                                ((new Random()).nextFloat() - 0.5F) * 0.02F + delta.x,
+                                ((new Random()).nextFloat() - 0.5F) * 0.02F,
+                                ((new Random()).nextFloat() - 0.5F) * 0.02F + delta.z);
                     }
                     else break;
                 }
